agentic_weather_chatbot/agentic-weather-chatbot/pyproject.toml
[project]
name = "agentic-weather-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "abdullahsheikh01", email = "162831270+abdullahsheikh01@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.19",
]

[project.scripts]
agentic-weather-chatbot = "agentic_weather_chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
